dev: add slow mechanism
(cherry picked from commit 74631f6de99dee046d4e14766d981937f03151d3)

+++ Assets/Scripts/bulletMovement.cs
@@ -8,7 +8,7 @@ public class bulletMovement : MonoBehaviour
     private Transform target;
     public float speed = 10f;
 
-    public int damage = 50;
+    public float damage = 50.0f;
 
     public float explodeRadius = 0f;
